refactor: move channels_v2/callbacks.py to channels/callbacks.py
