+++ apps/channels/channels_v2/connect_channel.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 import logging
 
+from apps.channels.callbacks import ChannelCallbacks
 from apps.channels.capabilities import ChannelCapabilities, PlatformConsentConfig
-from apps.channels.channels_v2.callbacks import ChannelCallbacks
 from apps.channels.channels_v2.channel_base import ChannelBase
 from apps.channels.channels_v2.pipeline import MessageProcessingContext
 from apps.channels.clients.connect_client import CommCareConnectClient
