+++ apps/channels/channels_v2/slack_channel.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import logging
 from typing import TYPE_CHECKING
 
-from apps.channels.channels_v2.callbacks import ChannelCallbacks
+from apps.channels.callbacks import ChannelCallbacks
 from apps.channels.channels_v2.channel_base import ChannelBase
 from apps.channels.const import MESSAGE_TYPES
 from apps.channels.datamodels import SlackMessage
