+++ apps/channels/channels_v2/whatsapp_channel.py
@@ -4,8 +4,8 @@ import logging
 from io import BytesIO
 from typing import TYPE_CHECKING
 
+from apps.channels.callbacks import ChannelCallbacks
 from apps.channels.capabilities import ChannelCapabilities
-from apps.channels.channels_v2.callbacks import ChannelCallbacks
 from apps.channels.channels_v2.channel_base import ChannelBase
 from apps.channels.channels_v2.stages.core import AttachmentHydrationStage
 from apps.channels.datamodels import WhatsAppMessage
