+++ apps/channels/tests/channels/concrete/test_web_channel.py
@@ -3,9 +3,9 @@ from unittest.mock import MagicMock, patch
 import pytest
 from django.http import Http404
 
+from apps.channels.callbacks import ChannelCallbacks
 from apps.channels.capabilities import ChannelCapabilities
 from apps.channels.channels_v2.api_channel import NoOpSender
-from apps.channels.channels_v2.callbacks import ChannelCallbacks
 from apps.channels.channels_v2.stages.core import (
     ConsentFlowStage,
     SessionResolutionStage,
