+++ apps/channels/tests/channels/conftest.py
@@ -3,8 +3,8 @@ from unittest.mock import MagicMock
 
 import pytest
 
+from apps.channels.callbacks import ChannelCallbacks
 from apps.channels.capabilities import ChannelCapabilities
-from apps.channels.channels_v2.callbacks import ChannelCallbacks
 from apps.channels.channels_v2.channel_base import ChannelBase
 from apps.channels.channels_v2.pipeline import MessageProcessingContext
 from apps.channels.const import MESSAGE_TYPES
